Added express-validator with his middleware

+++ backend/controllers/sauces.js
@@ -24,12 +24,19 @@ exports.modifySauce = (req, res, next) => {
     } : { ...req.body };
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            if (req.file != undefined) {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
                         .then(() => res.status(200).json({ message: 'Objet modifié !'}))
                         .catch(error => res.status(400).json({ error }));
-            })
+                })
+            }
+            else {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié !'}))
+                    .catch(error => res.status(400).json({ error }));
+            }
         })
         .catch(error => res.status(400).json({ error }));
 };
